refactor(domain) : WinningLotto의 내부 클래스 MatchResult를 독립 클래스로 분리
- MatchResult를 lotto.domain 패키지의 독립 클래스로 이동
- LottoResult 및 LottoCalculator에서도 재사용 가능하도록 구조 개선
- 의존성 단방향화 및 도메인 응집도 향상
