+++ src/main/java/lotto/domain/WinningLotto.java
@@ -39,22 +39,4 @@ public class WinningLotto {
 
         return new MatchResult((int) matchCount, bonusMatch);
     }
-
-    public static class MatchResult {
-        private final int matchCount;
-        private final boolean bonusMatch;
-
-        public MatchResult(int matchCount, boolean bonusMatch) {
-            this.matchCount = matchCount;
-            this.bonusMatch = bonusMatch;
-        }
-
-        public int getMatchCount() {
-            return matchCount;
-        }
-
-        public boolean isBonusMatch() {
-            return bonusMatch;
-        }
-    }
 }
